Move model enums to sensible locations
